Updated Fireball, Frostbolt, Rejuvenate, HorizCircleFrame, and VertCircleFrame. Fireball, Frostbolt, and Rejuvenate now have smaller effect trails to make sight easier for the player. HorizCircleFrame and VertCircleFrame now use a float radius value instead of an integer, allowing for smaller values to downsize the circle particle effects.

+++ src/main/java/us/fortherealm/plugin/skill/skills/formats/HorizCircleFrame.java
@@ -5,9 +5,9 @@ import org.bukkit.util.Vector;
 
 public class HorizCircleFrame implements ParticleFormat {
 
-    private int radius;
+    private float radius;
 
-    public HorizCircleFrame(int radius) {
+    public HorizCircleFrame(float radius) {
         this.radius = radius;
     }
 
@@ -24,7 +24,7 @@ public class HorizCircleFrame implements ParticleFormat {
         }
     }
 
-    public void setRadius(int radius) {
+    public void setRadius(float radius) {
         this.radius = radius;
     }
